refactor: migrate to standard Go project layout with cmd/ and internal/ directories
