+++ cmd/7dayspoll/main.go
@@ -1,17 +1,15 @@
 package main
 
 import (
-	"7DaysPoll/runtime"
+	"7DaysPoll/internal/bot"
 	"log"
 	"os"
 )
 
 func main() {
 	token := os.Getenv("DISCORD_BOT_TOKEN")
-	bot := runtime.NewBot(token)
-
-	err := bot.Run()
-	if err != nil {
+	b := bot.NewBot(token)
+	if err := b.Run(); err != nil {
 		log.Fatalln(err)
 	}
 }
