+++ internal/manage/delete.go
