+++ internal/manage/register.go
@@ -1,21 +1,20 @@
 package manage
 
 import (
-	"7DaysPoll/command"
 	"log"
 
 	"github.com/bwmarrin/discordgo"
+
+	"7DaysPoll/internal/poll"
 )
 
 func Register(session *discordgo.Session) error {
 	delete(session)
-	
 	log.Printf("Registering commands...\n")
-	_, err := session.ApplicationCommandCreate(session.State.User.ID, "", command.GetPollCommand())
+	_, err := session.ApplicationCommandCreate(session.State.User.ID, "", poll.GetPollCommand())
 	if err != nil {
 		return err
 	}
-
 	log.Printf("Command registration completed successfully.\n")
 	return nil
 }
