+++ internal/util/util.go
@@ -10,12 +10,10 @@ func GetTimeZone(lang string) (*time.Location, error) {
 	timezone := map[string]string{
 		"Japanese": "Asia/Tokyo",
 	}
-
 	tz, ok := timezone[lang]
 	if !ok {
 		return time.Local, nil
 	}
-
 	return time.LoadLocation(tz)
 }
 
@@ -24,12 +22,10 @@ func getWeekdays(lang discordgo.Locale) []string {
 		discordgo.EnglishUS: {"Sun", "Mon", "Tue", "Wed", "Thu", "Fri", "Sat"},
 		discordgo.Japanese:  {"日", "月", "火", "水", "木", "金", "土"},
 	}
-
 	weekdays, ok := localeWeekdays[lang]
 	if !ok {
 		return localeWeekdays[discordgo.EnglishUS]
 	}
-
 	return weekdays
 }
 
@@ -38,12 +34,10 @@ func getAbsence(lang discordgo.Locale) string {
 		discordgo.EnglishUS: "Absence",
 		discordgo.Japanese:  "欠席",
 	}
-
 	name, ok := absence[lang]
 	if !ok {
 		return absence[discordgo.EnglishUS]
 	}
-
 	return name
 }
